+++ File Copy Manager/ui/__init__.py
@@ -1,1 +1,0 @@
-"""UI package for File Copy Manager"""